fix that potentially solve nextjs issue
